# frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Java Interview Lab API",
    "description": "Java 后端面试学习系统接口",
    "version": "V0.1.1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/study/notes/{id}": {
      "put": {
        "tags": [
          "学习笔记"
        ],
        "summary": "保存笔记",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseNoteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/study/notes/{id}": {
      "put": {
        "tags": [
          "学习笔记"
        ],
        "summary": "保存笔记",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseNoteResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/questions/{id}": {
      "get": {
        "tags": [
          "题库"
        ],
        "summary": "查询题目详情",
        "description": "返回完整学习内容；不存在时返回 404。",
        "operationId": "detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseQuestionDetailResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "题库"
        ],
        "summary": "修改题目",
        "description": "version 过期时返回 409，调用方应先重新读取详情。",
        "operationId": "update_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseQuestionDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/study/notes": {
      "get": {
        "tags": [
          "学习笔记"
        ],
        "summary": "读取笔记",
        "operationId": "find",
        "parameters": [
          {
            "name": "targetType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "QUESTION",
                "TOPIC",
                "SCENARIO"
              ]
            }
          },
          {
            "name": "targetId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseNoteResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "学习笔记"
        ],
        "summary": "创建笔记",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseNoteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/study/notes": {
      "get": {
        "tags": [
          "学习笔记"
        ],
        "summary": "读取笔记",
        "operationId": "find_1",
        "parameters": [
          {
            "name": "targetType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "QUESTION",
                "TOPIC",
                "SCENARIO"
              ]
            }
          },
          {
            "name": "targetId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseNoteResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "学习笔记"
        ],
        "summary": "创建笔记",
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseNoteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/study/favorites/questions/{questionId}": {
      "post": {
        "tags": [
          "收藏"
        ],
        "summary": "收藏题目",
        "operationId": "favorite",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "收藏"
        ],
        "summary": "取消收藏题目",
        "operationId": "unfavorite",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/study/favorites/questions/{questionId}": {
      "post": {
        "tags": [
          "收藏"
        ],
        "summary": "收藏题目",
        "operationId": "favorite_1",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "收藏"
        ],
        "summary": "取消收藏题目",
        "operationId": "unfavorite_1",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/questions/{id}/answer-view": {
      "post": {
        "tags": [
          "V0.3 题目工作区"
        ],
        "summary": "显式披露参考答案",
        "description": "按 clientViewId 幂等记录披露行为并返回教学内容。",
        "operationId": "reveal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAnswerViewResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/questions": {
      "get": {
        "tags": [
          "题库"
        ],
        "summary": "分页查询题目",
        "description": "未传 page 和 pageSize 时分别默认为 1 和 20。",
        "operationId": "list_2",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/QuestionQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseQuestionSummaryResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "题库"
        ],
        "summary": "创建题目",
        "description": "专题和标签必须存在；同一种答案只能提交一次。",
        "operationId": "create_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseQuestionDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/questions/{id}": {
      "get": {
        "tags": [
          "V0.3 题目工作区"
        ],
        "summary": "读取练习元数据",
        "description": "不返回答案、讲解或追问参考答案。",
        "operationId": "metadata",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseQuestionWorkspaceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/questions/{id}/learning": {
      "get": {
        "tags": [
          "V0.3 题目工作区"
        ],
        "summary": "读取学习模式内容",
        "description": "用户主动进入学习模式后返回完整教学内容。",
        "operationId": "learning",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseQuestionLearningResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "traceId": {
            "type": "string"
          }
        }
      },
      "SaveNoteRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 0
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "content",
          "version"
        ]
      },
      "ApiResponseNoteResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/NoteResponse"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "NoteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "targetType": {
            "type": "string"
          },
          "targetId": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "QuestionAnswerRequest": {
        "type": "object",
        "properties": {
          "answerType": {
            "type": "string",
            "enum": [
              "QUICK_30S",
              "STANDARD",
              "DEEP"
            ]
          },
          "content": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 0
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "answerType",
          "content"
        ]
      },
      "QuestionFollowUpRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          },
          "referenceAnswer": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 0
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "title"
        ]
      },
      "QuestionUpdateRequest": {
        "type": "object",
        "description": "修改题目请求",
        "properties": {
          "topicId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          },
          "questionType": {
            "type": "string",
            "enum": [
              "KNOWLEDGE",
              "SCENARIO",
              "DESIGN_PATTERN"
            ]
          },
          "starLevel": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD"
            ]
          },
          "frequencyLevel": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "VERY_HIGH"
            ]
          },
          "originType": {
            "type": "string",
            "enum": [
              "BUILTIN",
              "USER",
              "IMPORTED",
              "FUTURE_AI"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ENABLED",
              "DISABLED"
            ]
          },
          "oneLiner": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 0
          },
          "plainExplanation": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 0
          },
          "designReason": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 0
          },
          "commonMistakes": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 0
          },
          "scorePoints": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 0
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "description": "详情接口返回的版本号",
            "example": 0,
            "minimum": 0
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionAnswerRequest"
            }
          },
          "followUps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionFollowUpRequest"
            }
          }
        },
        "required": [
          "difficulty",
          "frequencyLevel",
          "starLevel",
          "title",
          "topicId",
          "version"
        ]
      },
...
      "ApiResponseQuestionDetailResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/QuestionDetailResponse"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "QuestionDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "topicId": {
            "type": "integer",
            "format": "int64"
          },
          "topicName": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64"
          },
          "categoryName": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "questionType": {
            "type": "string"
          },
          "starLevel": {
            "type": "integer",
            "format": "int32"
          },
          "difficulty": {
            "type": "string"
          },
          "frequencyLevel": {
            "type": "string"
          },
          "originType": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "oneLiner": {
            "type": "string"
          },
          "plainExplanation": {
            "type": "string"
          },
          "designReason": {
            "type": "string"
          },
          "commonMistakes": {
            "type": "string"
          },
          "scorePoints": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagItem"
            }
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerItem"
            }
          },
          "followUps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FollowUpItem"
            }
          }
        }
      },
...
      "CreateNoteRequest": {
        "type": "object",
        "properties": {
          "targetType": {
            "type": "string",
            "enum": [
              "QUESTION",
              "TOPIC",
              "SCENARIO"
            ]
          },
          "targetId": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 0
          }
        },
        "required": [
          "content",
          "targetId",
          "targetType"
        ]
      },
...
      "AnswerViewRequest": {
        "type": "object",
        "properties": {
          "clientViewId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "clientViewId"
        ]
      },
      "AnswerViewResponse": {
        "type": "object",
        "properties": {
          "clientViewId": {
            "type": "string",
            "format": "uuid"
          },
          "duplicated": {
            "type": "boolean"
          },
          "viewedAt": {
            "type": "string",
            "format": "date-time"
          },
          "learning": {
            "$ref": "#/components/schemas/QuestionLearningResponse"
          }
        }
      },
      "ApiResponseAnswerViewResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/AnswerViewResponse"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "QuestionLearningResponse": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "integer",
            "format": "int64"
          },
          "contentMode": {
            "type": "string"
          },
          "oneLiner": {
            "type": "string"
          },
          "plainExplanation": {
            "type": "string"
          },
          "designReason": {
            "type": "string"
          },
          "commonMistakes": {
            "type": "string"
          },
          "scorePoints": {
            "type": "string"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerItem"
            }
          },
          "followUps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FollowUpItem"
            }
          }
        }
      },
...
      "QuestionCreateRequest": {
        "type": "object",
        "description": "新增题目请求",
        "properties": {
          "topicId": {
            "type": "integer",
            "format": "int64",
            "description": "所属专题 ID",
            "example": 1
          },
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          },
          "questionType": {
            "type": "string",
            "enum": [
              "KNOWLEDGE",
              "SCENARIO",
              "DESIGN_PATTERN"
            ]
          },
          "starLevel": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD"
            ]
          },
          "frequencyLevel": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "VERY_HIGH"
            ]
          },
          "originType": {
            "type": "string",
            "enum": [
              "BUILTIN",
              "USER",
              "IMPORTED",
              "FUTURE_AI"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ENABLED",
              "DISABLED"
            ]
          },
          "oneLiner": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 0
          },
          "plainExplanation": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 0
          },
          "designReason": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 0
          },
          "commonMistakes": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 0
          },
          "scorePoints": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 0
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionAnswerRequest"
            }
          },
          "followUps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionFollowUpRequest"
            }
          }
        },
        "required": [
          "difficulty",
          "frequencyLevel",
          "starLevel",
          "title",
          "topicId"
        ]
      },
...
      "ApiResponseQuestionWorkspaceResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/QuestionWorkspaceResponse"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseQuestionLearningResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/QuestionLearningResponse"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "QuestionQuery": {
        "type": "object",
        "properties": {
          "keyword": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64"
          },
          "topicId": {
            "type": "integer",
            "format": "int64"
          },
          "starLevel": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD"
            ]
          },
          "frequencyLevel": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "VERY_HIGH"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ENABLED",
              "DISABLED"
            ]
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 1
          }
        }
      },
      "ApiResponsePageResponseQuestionSummaryResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseQuestionSummaryResponse"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "PageResponseQuestionSummaryResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionSummaryResponse"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}